[PATCH] generic_serial.c portability fix
gs_wait_tx_flushed() should have its second argument (timeout) unsigned
long, not int.  One of the callers passes it MAX_SCHEDULE_TIMEOUT and
function itself compares argument with that value.  Since that's
LONG_MAX, we get breakage on all 64bit platforms.

+++ drivers/char/generic_serial.c
@@ -296,7 +296,7 @@ static int gs_real_chars_in_buffer(struct tty_struct *tty)
 }
 
 
-static int gs_wait_tx_flushed (void * ptr, int timeout) 
+static int gs_wait_tx_flushed (void * ptr, unsigned long timeout) 
 {
 	struct gs_port *port = ptr;
 	unsigned long end_jiffies;
